+++ src/test/java/de/alpharogroup/comparators/ComparatorFactoryTest.java
@@ -21,10 +21,11 @@
 package de.alpharogroup.comparators;
 
 import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertNotNull;
 
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
+import java.security.NoSuchAlgorithmException;
+import java.security.SecureRandom;
+import java.util.*;
 
 import org.testng.annotations.Test;
 
@@ -60,4 +61,65 @@ public class ComparatorFactoryTest
 		assertEquals(actual, expected);
 	}
 
+	/**
+	 * Test for method {@link ComparatorFactory#newMapValuesComparator(Map)}
+	 */
+	@Test
+	public void testNewMapValuesComparator() {
+
+		List<Integer> values;
+		List<Integer> actual;
+		List<Integer> expected;
+		// new scenario...
+		values = Lists.newArrayList(1, 2, 3, 4, 5);
+		Map<Integer, Integer> numberCounterMap;
+		numberCounterMap = newCounterMap(new HashMap<>(), values);
+		Comparator<Integer> integerComparator = ComparatorFactory.newMapValuesComparator(numberCounterMap);
+		assertNotNull(integerComparator);
+	}
+
+	/**
+	 * Test for method {@link ComparatorFactory#newRandomMapValuesComparator(Map, SecureRandom)}
+	 */
+	@Test
+	public void testNewRandomMapValuesComparator() throws NoSuchAlgorithmException {
+
+		List<Integer> values;
+		List<Integer> actual;
+		List<Integer> expected;
+		// new scenario...
+		values = Lists.newArrayList(1, 2, 3, 4, 5);
+		Map<Integer, Integer> numberCounterMap;
+		numberCounterMap = newCounterMap(new HashMap<>(), values);
+		Comparator<Integer> integerComparator = ComparatorFactory.newRandomMapValuesComparator(numberCounterMap, SecureRandom.getInstanceStrong());
+		assertNotNull(integerComparator);
+	}
+
+
+	/**
+	 * Factory method for create a map for and count elements of the given collection
+	 *
+	 * @param <K>
+	 *            the generic type of the elements
+	 * @param counterMap
+	 *            the counter Map
+	 * @param elements
+	 *            the elements
+	 * @return the new map ready to count elements
+	 */
+	private static <K> Map<K, Integer> newCounterMap(Map<K, Integer> counterMap,
+													Collection<K> elements)
+	{
+		Objects.requireNonNull(counterMap);
+		for (K element : elements)
+		{
+			if (counterMap.containsKey(element))
+			{
+				counterMap.merge(element, 1, Integer::sum);
+				continue;
+			}
+			counterMap.put(element, 0);
+		}
+		return counterMap;
+	}
 }